requirements.txt
APScheduler==3.9.1
astor==0.8.1
Babel==2.9.1
backports.zoneinfo==0.2.1
bce-python-sdk==0.8.64
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.1
colorama==0.4.4
cssselect==1.1.0
cssutils==2.4.0
cycler==0.11.0
Cython==0.29.28
decorator==5.1.1
dill==0.3.4
distlib==0.3.4
dnspython==2.2.1
et-xmlfile==1.1.0
filelock==3.6.0
flake8==4.0.1
Flask==2.1.1
Flask-APScheduler==1.12.3
Flask-Babel==2.0.0
fonttools==4.31.2
future==0.18.2
identify==2.4.12
idna==3.3
imageio==2.16.1
imgaug==0.4.0
importlib-metadata==4.2.0
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.2
lmdb==1.3.0
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.1
mccabe==0.6.1
multiprocess==0.70.12.2
multiprocessing-on-dill==3.5.0a4
networkx==2.6.3
nodeenv==1.6.0
numpy==1.21.5
opencv-contrib-python==4.4.0.46
opencv-python==4.5.5.64
openpyxl==3.0.9
packaging==21.3
paddleocr==2.4.0.4
paddlepaddle-gpu==2.2.2.post112
pandas==1.3.5
pathos==0.2.8
Pillow==9.1.0
platformdirs==2.5.1
pox==0.3.0
ppft==1.6.6.4
pre-commit==2.18.1
premailer==3.10.0
protobuf==3.20.0
pyclipper==1.3.0.post2
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
pyflakes==2.4.0
PyMuPDF==1.19.6
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
python-Levenshtein==0.12.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
pyzbar==0.1.9
requests==2.27.1
scikit-image==0.19.2
scipy==1.7.3
Shapely==1.8.1.post1
shellcheck-py==0.8.0.4
six==1.16.0
tifffile==2021.11.2
toml==0.10.2
tqdm==4.64.0
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
virtualenv==20.14.0
visualdl==2.2.3
waitress==2.1.1
Werkzeug==2.1.0
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
